+++ test/helper/Module/Test/Domain/TestDomainErrors.ts
@@ -6,6 +6,7 @@ export const TestDomainErrors = DefineDomainErrors(
     public entity_book: standard_entity_errors = 'not_found';
     public entity_author: standard_entity_errors | 'custom_1' = 'not_found';
     public entity_composite_author: standard_entity_errors | 'custom_1' = 'not_found';
+    public identity_primary_key_author: standard_entity_errors | 'custom_1' = 'not_found';
     public other_error = '';
   })(),
 );