+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Changed
+
+- Updated @hexancore/common to 0.13.3
+- RefId column default is varchar now.
+
+## Fixed
+
+- Fixed identity in UIntValueColumn
+
 ## [0.13.1] - 2024-02-02
 
 ## Added